Remove CopyrightUtils, export function directly
